+++ db/migrate/20170607150100_add_uploaded_files_count_to_sheets.rb
@@ -1,0 +1,6 @@
+class AddUploadedFilesCountToSheets < ActiveRecord::Migration[5.1]
+  def change
+    add_column :sheets, :uploaded_files_count, :integer
+    add_index :sheets, :uploaded_files_count
+  end
+end